fix: 9 bugs — PRESETS shared state restored, SourceEffect.speed handles Lottie, StitchedMedia.min_media_start returns float, add_group meaningful dimensions, is_audio guards empty StitchedMedia, Group validation by _type, overlap validation, clip timing validation, keyframe duration computation
- Restored PRESETS deepcopy to break shared mutable references
- SourceEffect.speed returns None for Lottie assets (no Speed param)
- StitchedMedia.min_media_start returns float (was int truncation)
- add_group/group_clips default widthAttr/heightAttr to 1920/1080
- is_audio guards against empty StitchedMedia medias list
- _check_group_required_fields identifies Groups by _type not tracks key
- validate() now checks for overlapping clips on each track
- New _check_clip_timing validates non-negative start, positive duration
- set_position/scale/rotation keyframes compute proper duration from consecutive times
- Removed invalid tests for transitions on internal Group tracks

2673 tests pass in 19s.

+++ src/camtasia/timeline/clips/stitched.py
@@ -40,9 +40,9 @@ class StitchedMedia(BaseClip):
         return len(self._data.get('medias', []))
 
     @property
-    def min_media_start(self) -> int:
+    def min_media_start(self) -> float:
         """Minimum media start offset in frames."""
-        return int(self._data.get('minMediaStart', 0))
+        return float(self._data.get('minMediaStart', 0))
 
     def clear_segments(self) -> None:
         """Remove all nested segments."""
